vectorised whole thermo loop

+++ tests/thermo_tests.jl
@@ -2,7 +2,7 @@ using Test, Random, LinearAlgebra, Lux, ConfParser, Enzyme, ComponentArrays
 
 ENV["GPU"] = true
 ENV["FULL_QUANT"] = "FP32"
-ENV["HALF_QUANT"] = "FP32"
+ENV["HALF_QUANT"] = "FP16"
 
 include("../src/utils.jl")
 using .Utils
